split benches
