+++ vortex-array/Cargo.toml
@@ -140,3 +140,8 @@ harness = false
 name = "expr_large_struct_pack"
 path = "benches/expr/large_struct_pack.rs"
 harness = false
+
+[[bench]]
+name = "chunked_dict_builder"
+harness = false
+required-features = ["test-harness"]
