+++ vortex-duckdb/src/exporter/mod.rs
@@ -24,8 +24,7 @@ use bitvec::view::BitView;
 pub use cache::ConversionCache;
 pub use decimal::precision_to_duckdb_storage_size;
 use itertools::Itertools;
-use vortex::arrays::DictVTable;
-use vortex::arrays::{ConstantVTable, StructArray, TemporalArray};
+use vortex::arrays::{ConstantVTable, DictVTable, StructArray, TemporalArray};
 use vortex::dtype::datetime::is_temporal_ext_type;
 use vortex::encodings::runend::RunEndVTable;
 use vortex::encodings::sequence::SequenceVTable;
