+++ vortex-duckdb/src/exporter/struct_.rs
@@ -54,8 +54,7 @@ mod tests {
     use std::ffi::CString;
 
     use vortex::IntoArray;
-    use vortex::arrays::DictArray;
-    use vortex::arrays::{ConstantArray, PrimitiveArray, VarBinViewArray};
+    use vortex::arrays::{ConstantArray, DictArray, PrimitiveArray, VarBinViewArray};
     use vortex::buffer::{BitBuffer, buffer};
     use vortex::error::{VortexExpect, VortexUnwrap};
     use vortex::validity::Validity;
